Move built formatters to static finals in the Formatter interface
