+++ library/src/main/java/com/flipkart/android/proteus/ProteusBuilder.java
@@ -118,94 +118,11 @@ public class ProteusBuilder {
         register("HorizontalProgressBar", new HorizontalProgressBarParser(), "ProgressBar");
 
 
-        Formatter NumberFormatter = new Formatter() {
 
-            private DecimalFormat formatter;
 
-            @Override
-            public String format(JsonElement elementValue) {
-                double valueAsNumber;
-                try {
-                    valueAsNumber = Double.parseDouble(elementValue.getAsString());
-                } catch (NumberFormatException e) {
-                    return elementValue.toString();
-                }
-                formatter = new DecimalFormat("#,###");
-                if (Build.VERSION.SDK_INT > Build.VERSION_CODES.GINGERBREAD) {
-                    formatter.setRoundingMode(RoundingMode.FLOOR);
-                }
-                formatter.setMinimumFractionDigits(0);
-                formatter.setMaximumFractionDigits(2);
-                return formatter.format(valueAsNumber);
-            }
-
-            @Override
-            public String getName() {
-                return "number";
-            }
-        };
-
-        Formatter DateFormatter = new Formatter() {
-
-            @SuppressLint("SimpleDateFormat")
-            private SimpleDateFormat from = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
-            @SuppressLint("SimpleDateFormat")
-            private SimpleDateFormat to = new SimpleDateFormat("d MMM, E");
-
-            @Override
-            public String format(JsonElement elementValue) {
-                try {
-                    // 2015-06-18 12:01:37
-                    Date date = from.parse(elementValue.getAsString());
-                    return to.format(date);
-                } catch (Exception e) {
-                    return elementValue.toString();
-                }
-            }
-
-            @Override
-            public String getName() {
-                return "date";
-            }
-        };
-
-        Formatter IndexFormatter = new Formatter() {
-            @Override
-            public String format(JsonElement elementValue) {
-                int valueAsNumber;
-                try {
-                    valueAsNumber = Integer.parseInt(elementValue.getAsString());
-                } catch (NumberFormatException e) {
-                    return elementValue.toString();
-                }
-                return String.valueOf(valueAsNumber + 1);
-            }
-
-            @Override
-            public String getName() {
-                return "index";
-            }
-        };
-
-        Formatter joinFormatter = new Formatter() {
-            @Override
-            public String format(JsonElement elementValue) {
-                if (elementValue.isJsonArray()) {
-                    return Utils.getStringFromArray(elementValue.getAsJsonArray(), ",");
-                } else {
-                    return elementValue.toString();
-                }
-            }
-
-            @Override
-            public String getName() {
-                return "join";
-            }
-        };
-
-        register(NumberFormatter);
-        register(DateFormatter);
-        register(IndexFormatter);
-        register(joinFormatter);
+        register(Formatter.DATE);
+        register(Formatter.INDEX);
+        register(Formatter.JOIN);
+        register(Formatter.NUMBER);
     }
 }
